Make autil_xalloc_(prepend/append) private
There is no reason to expose these functions now that a dedicated string
type and API are being introduced to the library.
These functions are still being used by the autil_bigint* functions, so
for now they will stay statically defined in the implementation section
of autil.h, but expect to see them remove (along with their tests) in
the future.