Remove some vpcmp functions
These functions not only had incorrect prototypes, but were also
basically useless if pass to a map, since the void ptr was pointing to
an object with a private definition.

If the user needs to use these with a map then they should probably just
define their own comparison functions for struct autil_TYPE* instead.