Rename autil_bstr* to autil_vstr*
The bstr was originally called a byte-string, but what I was actually
trying to go for was something closer to C++'s std::string_view, i.e. a
reference to some *other* not-owned-by-you byte slice.
The name vstr (view-of-string) was chosen to be more in-line with the
inteded use of this POD structure.