Remove fatal-error behavior in autil_bigint_shiftr
Prior to this change, shifting by more a bigint's bit count in
autil_bigint_shiftr would trigger a fatal error. This behavior stems
from the C undefined behavior from right-shifting greater than a type's
bit-width. However, for an abstract arbitrary-precision integer, right
shifting all bits of the integer *should* result in an integer with only
zeros, i.e. the integer zero. This commit alters the behavior of
autil_bigint_shiftr so that shifting by a number of bits greater than
the bit count of a particular bigint will set the value of that bigint
to zero.