Add clang-format guards against incorrectly formatting of macro
The autil_arr and autil_arr_const macros are incorrectly formatted by
clang-format-7 when these macros are used in function signatures,
presumably because clang-format thinks the text in question is a
function call rather than a function signature. This commit add
clang-format off/on guards to work around the clang-format bug.