Rename autil_bigint_new_utf8 to autil_bigint_new_text
Since this function does not actually parse non-ASCII characters (e.g.
U+2795 ➕) the name of this function was somewhat misleading. The word
"text" is still somewhat misleading since "text" does not specify
encoding, but it makes more sense to think of text as in "something that
could be in a string literal" which covers ASCII/UTF-8 and also makes
sense when paired with the comment that the function uses the same
grammar as autil_bigint_new_cstr.